beacon proxy structure
